Health bar updated
The health bar has been fixed. I have added the seed count an invisible wall to add to the level. The player receives damage from the enemy, after the player dies, the game will open a window with a game over text. It will then send you back to the menu.